requirements.txt
click==7.1.2
cytoolz==0.11.0
ecdsa==0.16.1
eth-hash==0.3.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
json-rpc==1.13.0
MarkupSafe==1.1.1
pycryptodome==3.10.1
rlp==2.0.1
six==1.16.0
toolz==0.11.1
Werkzeug==1.0.1
